Extracting persistence from the gui
This branch will extract the file writing from the gui. At this point,
though, I'm going to skip file writing (I know it works) in favor of
printing to console, so I can verify the output. Not sure if date
formatting will end up in splain or tagStore, both, or elsewhere.

Unfortunately, I forgot that the history I wanted to replay is on my work
computer. (This version won't open the text file. The latest version can.)
Will merge later.