Basic adjust, need a tui test version
Basic flag interpretation. I think. The flag only affects the sent time,
so there's a disconnect. I'll need to make a console version, albiet just
for testing. For the moment, I'm not actually sure it worked. If it did, I
need to take the adjust tag out of the user file. It's irrelevant,
historically. probably?