Last fix of this version: terse & undo
Final fix for this version of splaintime. Further changes require the
refactoring I've begun in the extract branch. Fixes some small frustrations.

Entering j8x will undo the last entry if a minute hasn't elapsed.
Ignores blank entries.

Improved saved text of adjusted time. Moves the adjust flag to rear
and prints just the hh:mm when that was done, rather than Date.toString()

Entering a tag with { will mark it as a subtask. The only significance
is surrounding the elaped time in parentheses.