Enables using a properties file for config
Checks for Splaintime.properties file for StPpreference values. Honors
values for finish fuse, initial tag, undo flag, and version. Disables
the config dialog. Notes this new file in readme, as well as the use
of positive minute values.

Disables the relabel flag, as that was removing the tag from history
rather than relabeling the active tag.