Enables restarting with a distinct tag immediately with !!nt
When I return from work, it's convenient to start my computer
and not note the time so I can adjust the time down. To be fair,
it may be late by a bit, depending on how long between when I enter
the door and walk to my pc.

SplainTime is aware of next time tag in StPreference. The value
is currently hardcoded as !!nt, rather than exposed via the properties
file. Perhaps later. ST tells TagStore to save this value until it
quicksaves. Provided the user hasn't entered another tag in the
meantime, TS will quicksave this string with an impossible time
to ensure a restarted TS uses this as a signal to save the value
immediately, rather than as a continuation of the previous tag. As
this is incompatible behavior, I've incremented the version.