Enables checking time since tag, but only in memory
TagStore has an in memory list of recorded Tags. No provision, yet,
to store this when SplainTime closes. StPreference is aware of a
time since flag. SplainTime is able to request a Duration from
TagStore of the rest of the input and reply that it doesn't remember
the tag or the hours and minutes since then. Adds this commented
flag to the example properties file.

Removes redundant compiler values from pom.xml; adds the
jar version to within the manifest, in case the filename changes. Tag
offers a toString() of when it is and its text.