brk chg 191123_1201 SplainTime handles noon and midnight
Splaintime.adjustToHhmmFormat() is now public scope to enable
testing. This is in preference to using robot or fake
ActionEvents and extracting the result from TagStore. It should
handle adjustments into yesterday, tomorrow, and during/to
noon. Provides tests for the relevant cases.

Tidy in (Test)ParsesInput removes unused import and a comment.i