feat i191125 initial H2, liquibase, slf4j usage
Initializing TagStore starts an embedded H2 db. Liquibase ensures
this has the expected schema. The files aren't in the right position
here, but will be in the deployed version. This currently
saves the db in the user's home folder as a dot file.
This includes an slf4j properties file, but liquibase comes with
a logback configuration, so it's not in use.