feat i191125 revised schema and identity file on startup
Ensures the splaintime db starts with a relevant schema and initial
data. This also ensures that it can identify the recording
device by writing an identity file in the local confiuration
folder. This replaces mac address and so on that have other
goals to achieve.

The liquibase file adds some relevant initial data, device types and
currently honored configuration values. While this could have used
type ids to identify the configurations, that will be more of a
user grouping convenience. Else, I'd be either canonicalizing the
type desc or the type id. Both of which st_configuration already
posess, more meaningfully. I may well ditch the type table, given
that I separated out the algorithm and device type tables.

0_initial_tables also uses separate constraints for the foreign
key relationships. Liquibase only provides update/delete
controls as part of the command, rather than in the column
constraint definition.

Removes the other two dates from category. The important aspect is
when it expires. While it's interesting that category x has existed
a while, it is better served by checking for a st_tag_category
record of the earliest date.