feat i191125 adding tag writes to db
Startup notes the current ZoneOffset. This isn't resilient across
political events that change the offset (ex leap second).

Adding a tag to the in-memory list of tags also writes a st_tag
record for person 1 and the current recording device. This
needs a refactor, as it's a direct insert, rather than with some
intentional dao.

Honors config flags to change time since and vet category
directives. Properties file controls the database filename.

Renames st_tag.recorded_when to st_tag.happened_when. The difference
with adjusted_when was insufficient to avoid confusion. This ensures
that happened_when indicates it is either the recording time,
or the adjusted time, when the person adjusts it. Adds a boolean
to indicate whether the adjustment was hhmm style (though that's
not currently honored). Adds seconds to the unhonored default
value for time format in st_configuation record.

Starts to record the changes for v4.3