brk chg i191125 adds tagId to (When)Tag; actually writes db tags
Previous sql for writing tags referred to a deprecated or misnamed
column. This uses the correct column names and reports whether
a user's time adjustment was absolute or as hour-minute format.
To accomodate finding the new tag in the db, this adds the tag
id to WhenTag and Tag (though Tag doesn't survive). Like as not,
this is a breaking change for anyone using those classes.