feat i200212 start of device sync protocol
Adds PeerSynchronizer to exchange records between sessions that
can access each other via a normal socket. This currently has
an over complicated exchange of recording device records, as I
thought it important to compare identities. Really, though, this
just needs a pigdin to inform the mapping. So, the next version
will just exchange lists of guids and send the records for each
unrecognized one.
Uses gson to exchange the records.