brk chg i200212 actual sync; avoid replacing device guid
PeerSynchronizer has the client and server just send all the
entire recording device records at one another. While this works
for the recording devices, I'll need to make some dto for tags,
as that could become a lot of traffic.
Describe the sync flag in canon properties, but doesn't currently honor
any changes to it. This is of the form '<>sync listen 37576' or
'<>sync send 27.13.311.1 37576'.

This turned up a problem for newly created databases, whereby st
rejected an existing guid, if it didn't match one it was aware of.
This is more an issue of my testing, as I don't expect someone to
move the device guid, if properly educated.
This is a breaking change, so I've incremented to v6.