fix i700101 default multirelease prep; remove version pofiles
Previously, this pom had multiple profiles to accommodate the
different toolchains (pre and post v9) as well as versions for
releasing. Thankfully, Lanterna published a template used here
whereby this compiles to v8 classes and puts module-info in
the multirelease folder.
Removes the now extraneous profiles in favor of always v8+mod
and one release profile, in addition to jooq. Also separates the
assembly plugin as a separate profile, as it's not kind to insist
that sonatype host the shadow-jar style bundle.
I'll need to infect my other projects from this template.